add metadata to CRDs, const for pool sizes
